# openapi/personnel-administration-api.yaml
openapi: 3.0.3
info:
  title: B4RRHH Personnel Administration API
  version: 0.1.0
paths:
...
  /rule-systems:
    post:
      operationId: createRuleSystem
      summary: Create rule system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleSystemRequest'
      responses:
        '201':
          description: Rule system created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSystemResponse'
    get:
      operationId: listRuleSystems
      summary: List rule systems
      responses:
        '200':
          description: Rule systems list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleSystemResponse'
  /rule-entity-types:
    post:
      operationId: createRuleEntityType
      summary: Create rule entity type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleEntityTypeRequest'
      responses:
        '201':
          description: Rule entity type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEntityTypeResponse'
    get:
      operationId: listRuleEntityTypes
      summary: List rule entity types
      responses:
        '200':
          description: Rule entity types list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleEntityTypeResponse'
  /rule-entity-types/{code}:
    get:
      operationId: getRuleEntityTypeByCode
      summary: Get rule entity type by code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        '200':
          description: Rule entity type found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEntityTypeResponse'
        '404':
          description: Rule entity type not found
  /rule-entities:
    post:
      operationId: createRuleEntity
      summary: Create rule entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleEntityRequest'
      responses:
        '201':
          description: Rule entity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEntityResponse'
    get:
      operationId: listRuleEntities
      summary: List rule entities
      parameters:
        - in: query
          name: ruleSystemCode
          required: false
          schema:
            type: string
            maxLength: 5
        - in: query
          name: ruleEntityTypeCode
          required: false
          schema:
            type: string
            maxLength: 30
        - in: query
          name: code
          required: false
          schema:
            type: string
            maxLength: 30
        - in: query
          name: active
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Rule entities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleEntityResponse'
  /rule-entities/{id}:
    get:
      operationId: getRuleEntityById
      summary: Get rule entity by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rule entity found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEntityResponse'
        '404':
          description: Rule entity not found
  /rule-entities/search:
    get:
      operationId: searchRuleEntityByBusinessKey
      summary: Search rule entity by business key
      parameters:
        - in: query
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: query
          name: entityTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: query
          name: code
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        '200':
          description: Rule entity found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEntityResponse'
        '404':
          description: Rule entity not found
  /rule-systems/{code}:
    get:
      operationId: getRuleSystemByCode
      summary: Get rule system by code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            maxLength: 5
      responses:
        '200':
          description: Rule system found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSystemResponse'
        '404':
          description: Rule system not found
components:
  schemas:
...
    CreateRuleSystemRequest:
      type: object
      required:
        - code
        - name
        - countryCode
      properties:
        code:
          type: string
          maxLength: 5
        name:
          type: string
          maxLength: 100
        countryCode:
          type: string
          minLength: 3
          maxLength: 3
    RuleSystemResponse:
      type: object
      required:
        - id
        - code
        - name
        - countryCode
        - active
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        countryCode:
          type: string
        active:
          type: boolean
    CreateRuleEntityTypeRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
    RuleEntityTypeResponse:
      type: object
      required:
        - id
        - code
        - name
        - active
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        active:
          type: boolean
    CreateRuleEntityRequest:
      type: object
      required:
        - ruleSystemCode
        - ruleEntityTypeCode
        - code
        - name
        - startDate
      properties:
        ruleSystemCode:
          type: string
          maxLength: 5
        ruleEntityTypeCode:
          type: string
          maxLength: 30
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
    RuleEntityResponse:
      type: object
      required:
        - id
        - ruleSystemCode
        - ruleEntityTypeCode
        - code
        - name
        - active
        - startDate
      properties:
        id:
          type: integer
          format: int64
        ruleSystemCode:
          type: string
        ruleEntityTypeCode:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        active:
          type: boolean
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true